+++ CHANGELOG.md
@@ -4,7 +4,11 @@ To install or update LODA, please follow the [installation instructions](https:/
 
 ### Features
 
-* Added `fac` operation for falling and rising factorial
+* Added `fac` operation for falling/rising factorial
+
+# Enhancements
+
+* Use two-complement representation for bitwise operations
 
 ## v25.7.14