+++ CHANGELOG.md
@@ -2,6 +2,8 @@ To install or update LODA, please follow the [installation instructions](https:/
 
 ## [Unreleased]
 
+## v25.7.24
+
 ### Features
 
 * Added `fac` operation for falling/rising factorial